beta-classifier/pyproject.toml
[project]
name = "beta-classifier"
version = "0.1.0"
description = "Climbing grade classifier using move sequences from beta solver"
authors = [{name = "Moonboard Grader Team"}]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.6.0",
    "pyyaml>=6.0.2",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "tensorboard>=2.18.0",
    "tqdm>=4.67.0",
    "moonboard_core",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv.sources]
moonboard_core = { path = "../moonboard_core", editable = true }

